requirements.txt
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Telegram bot
aiogram==3.15.0

# AI / STT
openai==1.57.2
faster-whisper==1.1.0

# HTTP & async utilities
httpx==0.28.1
aiofiles==24.1.0
sse-starlette==2.1.3
starlette>=0.40.0,<0.42.0

# Logging & dev
python-dotenv==1.0.1

redis>=5.0.0          # aiogram RedisStorage for persistent FSM
apscheduler>=3.10.0   # scheduled Telegram reports (daily/weekly/monthly)
openpyxl>=3.1.0       # Excel (.xlsx) export
